udpate torch compile tests
