+++ transformer_nuggets/quant/__init__.py
@@ -1,2 +1,3 @@
-from transformer_nuggets.quant.qlora import get_block_absmax, linear_nf4, NF4Tensor
+from transformer_nuggets.quant.nf4_tensor import get_block_absmax, NF4Tensor
+from transformer_nuggets.quant.qlora import linear_nf4
 from transformer_nuggets.quant.qlora_debug import NF4TensorDebug
